Payment Sayfası ve businessda kurallar yazıldı

+++ Entities/Concretes/PayInvoicePerson.cs
@@ -11,7 +11,7 @@ namespace Entities.Concretes
     {
         public int InvoiceId { get; set; }
         public int PersonId { get; set; }
-        public double PayingAmount { get; set; }
+        public decimal PayingAmount { get; set; }
         public DateTime DueDate { get; set; } = DateTime.Now;
     }
 }
